config/ingress: Remove broken hostname validation
The hostname validation for the custom route configuration API prohibits
valid hostnames, such as hostnames with TLDs that include digits.  This API
was added in OpenShift 4.8, and we cannot tighten OpenAPI validation of a
shipped API.  Moreover, allowing some invalid values while rejecting other
invalid values in the API is more confusing than consistently allowing
invalid values in the API and then handling invalid values in the component
that owns the route.  Thus this commit removes the validation entirely.

This commit fixes bug 2039256.

https://bugzilla.redhat.com/show_bug.cgi?id=2039256

* config/v1/types_ingress.go (Hostname): Remove validation.
* config/v1/0000_10_config-operator_01_ingress.crd.yaml: Regenerate.