`DefaultPlacement` allow empty enum value and remove default
The `DefaultPlacement` config status field should allow the `""` value
to be used during an Update for this subresource. This should allow the
status subresource to be updated even when `DefaultPlacement` has an
empty value.

Also removed the `kubebuilder:default:="Workers"` specifier because the
controller already regards an empty value as `"Workers"`, there's no
need for it to be explicitly set when omitted.