Allow 0 values for host network IngressControllers
Allow the host network IngressController ports to have 0 as a valid
value for when the user wants to specify "no-preference". Also added the
`+optional` tags for the ports and `omitempty` which will ensure that
when the fields are not specified at all the serialized resource will
not contain that field.