vSphere: add zonal spec to infrastructure
This PR implements the required fields necessary to add zonal topology
for vSphere-based OpenShift clusters as a Tech Preview feature.

It defines the vCenter based objects: datacenter, cluster, network,
datastore, resource pool, folder and tag names that are required to
provision and manage the arbitrary nature of vSphere and
region and zone topology.