config/v1/types_cluster_version: Add architecture fields
As described in [enhancement], add architecture fields
to ClusterVersionSpec and to ClusterVersionStatus. The
former indicates the desired architecture and the later
indicates the current architecture. A cluster's
architecture can be either homogeneous or heterogeneous.
A single valid architecture value, e.g. "amd64",
indicates an homogeneous cluster and "multi" indicates
a heterogeneous cluster.

[enhancement]: https://github.com/openshift/enhancements/blob/836516786ce47bf74a8a3dfe8a2fa7f1143e94c9/enhancements/multi-arch/heterogeneous-architecture-clusters.md