config/v1: Change ClusterVersion capability from 'console' to 'Console'
The lowercase form just landed via 9db3a0ce01 (CONSOLE-3221: Add
console capability (#1232), 2022-08-01), and that matched the
lower-case precedent which we had been using for these enumerated
capabilities.  But the Kube-API preference is for PascalCase [1], so
this commit pivots to that form.  Because we are so close on the heels
of the original change, we don't need to preserve any references to
the outgoing form.  Adjusting the 4.11 references like
openshift-samples will take more dancing, and can happen in follow-up
work.

Generated with:

  $ sed -i 's|\(capability[.]openshift[.]io/name\): console|\1: Console|' $(git grep -l 'capability.openshift.io/name: console')
  $ sed -i 's/console/Console/' config/v1/*version*

And then backing out one godocs hit that the above made:

  $ sed -i 's/the web Console/the web console/' config/v1/*version*

[1]: https://github.com/openshift/enhancements/blame/b3f5408be4505e37c0bdb7f8b2630a92a5d01457/dev-guide/api-conventions.md#L195