route/v1: Fix CRD to allow path for non-TLS routes
Fix the Route CRD validation for allow routes with null spec.tls and
nonempty spec.path.

This commit is related to USHIFT-333.

https://issues.redhat.com/browse/USHIFT-333

* route/v1/route.crd.yaml-patch: Add a case to allow spec.path to be
nonempty when spec.tls is null.  The existing case to allow spec.path when
spec.tls.termination is not "passthrough" is insufficient when spec.tls is
not specified at all.
* route/v1/route.crd.yaml: Regenerate.
* route/v1/test-route-validation.sh: Add test cases "non-TLS with nonempty
path" and "edge-terminated with nonempty path".