Speaker ASN is optional
The user can specify what ASN the speakers should have; or
baremetal-runtimecfg will use a predefined one.

Rationale: The BGP config is bound to a failure-domain. Which is fair:
we expect each failure domain to have a fixed set of (leaf) routers to
advertise the API VIP to. However, what happens when we fit more than
one master in a failure domain? Each should get a distinct ASN, since
they would be talking to the same uplink router. So they can share the
rest of the config, but not the ASN. Here I propose that when ASN is not
set, we hardcode arbitrary ASN values in the Private Use range (ie
4200000000-4294967294).