pin default beta enable to 1.25
All new kube beta features will be disabled in a standard deployment by default.
A team may explicitly opt-in to enable a particular beta feature by
adding it to the list, signing the team up for support and signing up to
keep the feature backwards compatible and upgradeable even though
kubernetes does not guarantee this.

All new kube beta features will be enabled in TechPreview.
A team may only disable beta features if there is a severe, intractable
bug or if the kubernetes project plans to keep a feature disableable
forever.  Keep in mind that kube has a stated goal for all featuregates
to become force-enabled with no residual code to allow long term
disablement.