MGMT-13627: Add ConfidentialVM and TrustedLaunch options to AzureMachineProviderSpec
This change introduces the security profile to the OS disk parameters
and adds the SecurityType and UEFISettings sections to the VM's security
profile.

The SecurityType defines whether the VM is a Confidential or a Trusted Launch
VM. This field should be set to one of the two options, in order to enable
the UEFISettings section.

The UEFISettings fields include the VirtualizedTrustedPlatformModule and
SecureBoot fields, which can be set to either Enabled or Disabled.

The OS disk security profile includes the SecurityEncryptionType and the
DiskEncryptionSet fields. The SecurityEncryptionType can only be used when
the VM is a Confidential one. When SecurityEncryptionType is set, vTPM
should be enabled and the VM security profile's SecurityType should also
be set to ConfidentialVM.

Possible values for the SecurityEncryptionType are:
- DiskWithVMGuestState, OS disk encryption before VM deployment that
  uses platform-managed keys (PMK) or a customer-managed key (CMK)
- VMGuestStateOnly, no OS disk confidential encryption

When the SecurityEncryptionType is set to DiskWithVMGuestState, the
DiskEncryptionSet can be specified to allow for customer-managed keys
to be used to encrypt the OS disk and the VMGuestState blob.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>