OPNET-263: Enable CloudDualStackNodeIPs feature gate
In order to allow passing dual-stack IPs to the `--node-ip` when using
cloud provider, we are enabling the CloudDualStackNodeIPs feature gate.

This affects only behaviour of deployments that want to pass 2 IP
addresses as `--node-ip` together with `--cloud-provider`. Before K8s
1.27 this was not possible (you can either have dual-stack Node IPs or
use a cloud provider, but not both).

Upstream KEP https://github.com/kubernetes/enhancements/issues/3705
adds ability to pass exactly 2 IPs as `--node-ip` also when using cloud
provider, but given its alpha stage, feature gate is required here.

Enablement of this feature gate is not affecting in any way single-stack
deployments as well as dual-stack deployments with no cloud provider.

Enablement of this feature gate is not affecting in any way single-stack
deployments that use cloud provider.

Enablement of this feature gate is not changing the behaviour of any
configuration that is working today (i.e. before K8s 1.27).

Relates-to: OPNET-13
Relates-to: OPNET-230
Fixes: OPNET-263