Add config knob for enabling ip forwarding
With https://github.com/ovn-org/ovn-kubernetes/pull/3374 and
https://github.com/openshift/machine-config-operator/pull/3676
we stopped enabling IP forwarding globally in the OCP host. It is now
only enabled on interfaces that OVNK manages. Additionally iptables
rules are put in place to limit what traffic is forwarded. This was to
resolve an issue where our node by default acts as a router:
https://issues.redhat.com/browse/OCPBUGS-3176

With these changes, by default in OCP we will no longer forward IP
traffic. Most users should welcome this change as it means the OCP node
will not be forwarding traffic unintentionally on their network.
However, in case there is some corner case where a user expects traffic
to be forwarded we will expose this new config knob to allow them to
go back to the old behavior.

Signed-off-by: Tim Rozet <trozet@redhat.com>