*: Promote TechPreviewNoUpgrade resources into CustomNoUpgrade
CustomNoUpgrade is even-more-cutting-edge than TechPreviewNoUpgrade
[1].  Before this commit, custom resources with properties with:

  +openshift:enable:FeatureSets=TechPreviewNoUpgrade

were being sharded into Default and TechPreviewNoUpgrade flavors,
which left CustomNoUpgrade clusters with neither shard.  With this
commit, the CustomNoUpgrade cluster will get the same CRD as the
TechPreviewNoUpgrade cluster.  It's not entirely clear to me how folks
would annotate a CustomNoUpgrade-only resource, because:

  +openshift:enable:FeatureSets=CustomNoUpgrade

might not generate a:

  release.openshift.io/feature-set: TechPreviewNoUpgrade,Default

manifest for the less-cutting-edge clusters.  But we can work that out
in future work, and get this commit in now to get CustomNoUpgrade back
off the ground.

Generated with:

  $ sed -i 's/openshift:enable:FeatureSets=TechPreviewNoUpgrade/openshift:enable:FeatureSets=CustomNoUpgrade,TechPreviewNoUpgrade/' $(git grep -l openshift:enable:FeatureSets=)

[1]: https://github.com/openshift/api/pull/370#issuecomment-510632939