Add an owning-component annotation.
Given the large number of resources we have, it is useful to have a
mechanism to determine which component to contact when a particular one
needs attention.  The first direct usage will be for recording who is
responsible for in-cluster secrets and configmaps that contain certs,
keys, and CA bundles.