Add configuration options for custom DNS work.
The work relates to the enhancement https://github.com/openshift/enhancements/pull/1468. The AWS, Azure, and GCP platform status structs are updated to include custom DNS options. The internal and external load balancer ip addresses as well as the types of dns records make up the base of the data.y