OCPBUGS-20164: Capability Annotation for Build CRD
Add missing `capability.openshift.io/name` annotation to the Build cluster
configuration CRD. This will ensure that the config object for the Build
sub-system is not created when the Build capability is not enabled.
Also annotate the empty-resources Build CRD so that it is linked with the
appropriate cluster capability in the payload empty-resources.