MCO: make certificate dates optional
So back in 4.14 we had to change the types of our cert observability
certificate dates from strings to metav1.Time. This resulted in some API
breakages, and in our haste for 4.14 we removed those fields but left
the rest of the object.

The fields have now been added back in 4.15, and we are hitting some
timing issues between when the "new" CRD gets applied, and when the
"old" pods get replaced.

The errors are unpleasant and are blocking CI, so we're going to make
these fields optional for 4.15 and then lock them back down to required
in 4.16. We will not have this issue during the 4.15->4.16 transition
because the MCO has full control of these fields and will ensure they
are populated in 4.15.