Consistent minTLSVersion in IngressController CRD.
The minTLSVersion sample values proposed in this CRD's documentation are leading to an error because the allowed enum values differ from the syntax of the config samples.

Symptoms observed on the operator upon applying an IngressController with the suggested values were: 'error: ingresscontrollers.operator.openshift.io "" is invalid'
see also https://github.com/openshift/cluster-ingress-operator/pull/994

Fixed the documentation for minTLSVersion values:
  $ find . -not -path "./.git/*" -type f -exec grep -Iq . {} \; -print0 | xargs -I {} -0 sed -Ei 's/minTLSVersion: TLSv1.([0-3])/minTLSVersion: VersionTLS1\1/g' {}

Replaced plural form for tls versions in the documentation:
  $ find . -not -path "./.git/*" -type f -exec grep -Iq . {} \; -print0 | xargs -I {} -0 sed -i 's/MinTLSVersions is/minTLSVersion is/g' {}

Did not replace "MinTLSVersion is the minimum TLS version supported" which uses an uppercase "M", as all other options are also starting with an uppercase letter.

Not sure whether it's correct to also update the generated files - let me know if I should remove changes that are not necessary.

Documentation reference for the TLS version naming: https://pkg.go.dev/crypto/tls#pkg-constants