OCPBUGS-24573: Generate Tech Preview version of ControllerConfig CRD
ControllerConfig CRD consumes fields that are part of the Infrastructure
CRD, e.g. in the part of the Platform Spec. As those contain different
fields depending whether Default, TechPreviewNoUpgrade or
CustomNoUpgrade is used, we also need to have ControllerConfig in
respective versions.

This PR creates TechPreviewNoUpgrade and CustomNoUpgrade next to the
already existing Default.

Fixes: OCPBUGS-24573