infra: change AWS ServiceEndpoints from map to atomic
The AWSServiceEndpoint struct has tags defined in a wrong way, i.e. it
is missing the `+kubebuilder:validation:Required` in one of its fields.
Because of that, AWSPlatformSpec cannot define ServiceEndpoints as a map
where names are used as keys because that would mean we have an optional
field used as a key in the map. Illegal.

It is not trivial to modify AWSServiceEndpoint struct to add a tag
making the Name required, so instead we are moving back ServiceEndpoints
inside AWSPlatformSpec to the atomic type.