MachineOSConfig and MachineOSBuild API
machineOSBuild and Image are meant to represent the two components of an on cluster build, making this commit the
user-facing API for on cluster builds. MOSBuilds comprise of the actual processes of kicking off an monitoring a build.

A MachineOSConfig contains user options like the push/pull spec for registries as well as the custom content in a dockerfile.  a MOSImage
is the result of a build. the Image objects are intended to be trackable and easily related to an MCP. The goal here is to remove much of the
split state reporting for on cluster builds and consolidate where the user sets config options.

Signed-off-by: Charlie Doern <cdoern@redhat.com>