config/ingress: Make Hostname godoc user-friendly
Add the necessary kubebuilder syntax to the godoc for the ingress config
API's Hostname type in order to hide developer notes from the generated
CRD's description for the status.componentRoutes.currentHostnames field.
Also, add a user-oriented description of the field with a reference to the
relevant RFC.

Before this commit, the oc explain output for the API field described it in
terms of its Go type, kubebuilder syntax, and regular-expression syntax for
validation.  After this commit, the oc explain output for the API field
describes it in terms that are relevant to the end-user.

The commit that added the validation did not mention *which* RFC it was
enforcing.  The closest RFC seems to be RFC-1123 in that the validation
allows the host name to start with a decimal digit, although the validation
seems a little looser than RFC-1123 in that allows arbitrary Unicode
letters and symbols with \p{L} and \p{S}.  The intent seems to be that the
end-user should use RFC-1123 compliant host names, and validation is looser
only as necessary for backwards-compatibility.

This commit fixes bug OCPBUGS-31058.

https://issues.redhat.com/browse/OCPBUGS-31058

Follow-up to commit 1fec415423985530a8925a5fd8c87e1741d8c2fb.

* config/v1/types_ingress.go: Update the Hostname type godoc.
* config/v1/0000_10_config-operator_01_ingress.crd.yaml:
* config/v1/zz_generated.featuregated-crd-manifests/ingresses.config.openshift.io/AAA_ungated.yaml:
* payload-manifests/crds/0000_10_config-operator_01_ingress.crd.yaml: Regenerate.