config/v1/types_infrastructure: change set to atomic for networks
In order to handle ownership of network addresses in a more elegant way,
we are switching list type from `set` to `atomic`. Thanks to this the
whole list will be managed as one, instead of different owners for every
entry on the list.

We are adding CEL validation to fields in PlatformStatus for the
consistency. As this field has already been validated by o/installer,
this should not affect CRs with PlatformStatus already populated.