Add FeatureGateAwareXValidation marker to eipAllocations CEL
The CEL validation added to `LoadBalancerStrategy` in #1826
causes `undefined field` errors if `SetEIPForNLBIngressController`
is not enabled and the scope is not internal.

The CEL validation added to `NetworkLoadBalancerParameters` also in #1826,
causes `undefined field` errors if `SetEIPForNLBIngressController` is
enabled without `IngressControllerLBSubnetsAWS`, or visa versa. Since
`FeatureGateAwareXValidation` doesn't support using AND for multiple
feature gates, we will resolve this by adding `SetEIPForNLBIngressController`,
the feature that originally introduced this CEL.