status: make namespaced
Initial draft used a non-namespaced API, because in standalone clusters
it would be a singleton scoped to the cluster, semantically. We expect
the concept to work with HCP too though. There the resource would need
to live in the management cluster apiserver and would be scoped to a
*hosted* cluster, not the management cluster (management cluster would
need one too, though). Therefore, we would need a namespaced,
functionally equivalent CRD for HCP, which seems like higher effort and
complexity than the minor awkwardness of using a namespaced resource in
standalone clusters.