config/v1/types_cluster_version: Add v4.17 and v4.18 capability sets
These are both synonyms for v4.16, because there have not been
capability-recommendation changes since that release.  4.17 is about
to go Generally Available, which means v4.17 will always be a synonym
for v4.16.  v4.18 may or may not evolve before 4.18 goes GA.  I think
defining synonym capability sets is busywork that takes developer time
without adding customer value, but [1] points out that so far every
4.y release has had a v4.y capability set string associated with it,
and folks might find adjustments to that pattern surprising.  I'm
declaring sets for our next two 4.y here to continue the existing
set-for-each-4.y pattern (and breaking the
set-for-each-4.y-with-a-different-default pattern) to unblock 4.17's
GA, but we may revisit this approach in the future with more lead
time.

[1]: https://issues.redhat.com/browse/OCPBUGS-41111