config/v1/types_cluster_version: Add v4.17 capability set
This is a synonym for v4.16, because there have not been
capability-recommendation changes since that release.  4.17 is about
to go Generally Available, which means v4.17 will always be a synonym
for v4.16.  I think defining synonym capability sets is busywork that
takes developer time without adding customer value, but [1] points out
that so far every 4.y release has had a v4.y capability set string
associated with it, and folks might find adjustments to that pattern
surprising.  I'm declaring the v4.17 set to continue the existing
set-for-each-4.y pattern (and breaking the
set-for-each-4.y-with-a-different-default pattern) to unblock 4.17's
GA, but we may revisit this approach in the future with more lead
time.  This commit backports (without v4.18) [2] to release-4.17.

[1]: https://issues.redhat.com/browse/OCPBUGS-41111
[2]: https://github.com/openshift/api/pull/2022