fix: added missing open-api gen tags to api packages
Added the missing tag to:
- cloudnetwork/v1
- imageregistry/v1
- machineconfiguration/v1
- networkoperator/v1

These tags are necessary so that the generated zz_generated.openapi.go
and openapi.json are complete.

In the current state, these files are missing some definition even though they
do have references to these missing definitions (e.g. EgressRouter has a reference
to EgressRouterStatus which is not present)