OCPBUGS-39148: Add Feature Gate AND on NetworkLoadBalancer CEL
Previously, the AWSNetworkLoadBalancerParameters struct had CEL that
referenced both Subnets and EIPAllocations, but only was gated on the
EIPAllocations feature gate. This means if the subnets feature gate was
ever disabled, then the IngressController CRD would be invalid because
the CEL would still be present, causing an error when installing the
CRD.

We are now able to add "AND" logic to the FeatureGateAwareXValidation
tag which enables us to feature gate the CEL on both Subnets and
EIPAllocations feature gates.