MachineOSBuild: make ObjectReference fields optional
These seem to have been made required during alpha API as it was copied
from an existing reference. These do not need to be required and it
helps with testing to set them as optional. This also makes it closer to
the corev1 ObjectReference type.

This should be also safe from a compatibility perspective, and even if
not, this API has not yet been used, so there isn't any dependency on
it.