Support SEV_SNP confidential type selection on GCP
GCP compute API permits not only specifying whether a machine has to be
confidential or not, but also determining the technology backing that
up see [0]. This commit adds an api parameter into the
GCPMachineProviderSpec to support that.

That way, we should be able to configure SEV_SNP machines as well as SEV
machines as before.

[0] https://cloud.google.com/confidential-computing/confidential-vm/docs/supported-configurations