Support SEV_SNP confidential type selection on GCP
GCP compute API permits not only specifying whether a machine has to be
confidential or not, but also determining the technology backing that
up see [0]. This commit adds the possibility of not only Enabling or
Disabling confidential computing nodes, but also choosing the type of
confidential computing instances that are to be configured.

That way, configuring SEV_SNP machines should be possible as well as SEV
machines as previously.

[0] https://cloud.google.com/confidential-computing/confidential-vm/docs/supported-configurations