status: break workerpool and controlplane structs
I was running into trouble with CRD validation because of CEL budgets.
The way the API is structured makes cardinatilities high because we
need to allow high amouts of insights because of nodes. Then the
hypothetical worst case leads to 32 worker pools, each with 16 informers
reporting 1024 insights, all of which are `ClusterVersion` status insights
which have a CEL that needs validation. Such scenario is purely
hypothetical but compliance to budgets is enforced by the tools.

The cardinality is so high because of the top-level worker pool list (32)
so the problem can be solved by separating what insights can be present
in which section of the API, which is what this commit does.

Alternatively, we could flatten the `informers` layer and that way we
would get rid of the `16` factor, and we would be able to limit the amount
of insights reported by all informers that way. The informer source is
not that important for API consumers but would make the API harder to
work with on the producer side.