Validate aws resource tag keys with aws prefix & immutable
AWS docs indicate that tag keys cannot be prefix with aws:. See:
https://docs.aws.amazon.com/directoryservice/latest/devguide/API_Tag.html

Using this key prefix leads to an AWS API error indicating the prefix
is reserved for AWS system usage. This commit adds API validation
and ratecheting tests, as the key was previously allowed.

Adds validation to enforce immutability on AWS resourcetags, in
the same manner as Azure & GCP. Updating resourcetags post-install
is not supported.