OTA-1339: Add `UpdateStatusAPI` FG for DevPreview
OTA originally created and used the UpgradeStatus legacy feature gate for the related functionality, enabled in TechPreview. Moving the functionality into the cluster, exposed by an API proved to be more complex and experimental than thought, so it was decided to create and use a new feature gate that is only enabled in DevPreview.