operator/v1alpha1: Do not create ClusterVersionOperator CRD in HyperShift
The CVO in HyperShift will not use the ClusterVersionOperator resource in
the hosted cluster. Thus, there is no need to create the CRD.

The CVO will still utilize the feature gate in HyperShift, as there will be a
new internal logic to read and parse a configuration file provided by a
HyperShift component. However, the CRD itself is not needed.